+++ src/main/java/main/lab1/services/implementation/UserServiceImpl.java
@@ -1,10 +1,11 @@
-package main.lab1.services;
+package main.lab1.services.implementation;
 
 import java.util.List;
 import main.lab1.model.User;
 import main.lab1.exceptions.DuplicateResourceException;
 import main.lab1.exceptions.UserNotFoundException;
 import main.lab1.repos.UserRepository;
+import main.lab1.services.UserService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,10 +21,6 @@ public class UserServiceImpl implements UserService {
 
     @Transactional
     public User createUser(User newUser) {
-        // Проверяем существование пользователя по ID
-        if (userRepository.existsById(newUser.getUserId())) {
-            throw new DuplicateResourceException("User with id " + newUser.getUserId() + " already exists");
-        }
         if (userRepository.existsByEmail(newUser.getEmail()))
         {
             throw new DuplicateResourceException("User with email " + newUser.getEmail() + " already exists");
